Add SodaException; Arrays vs objects handled diff
Add a new SodaException that will be called whenever there's an issue with a Soda request. An HttpException should only be thrown when Soda doesn't return any JSON and it's an actual server issue.
The returned JSON will always be handled as an associative array internally but it will be returned as an object if the user requests it.